// NSDL_FVCI_UAT_RELEASE/13-03-2025/Website Deployment/nsdlfpi-development/src/swagger/model/draftFvicRegistrationDetailsDto.ts
/**
 * NSDL API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface DraftFvicRegistrationDetailsDto { 
    id?: number;
    fvciApplicationId?: string;
    typeOfApplicant?: string;
    otherTypeOfApplicant?: string;
    typeOfEntity?: string;
    isProvidedFactaCrsProvided?: boolean;
    isComingFromGlobalCustodian?: boolean;
    ddpName?: string;
    ddpRegistrationNumber?: string;
    custodianName?: string;
    custodianRegistrationNumber?: string;
    dpName?: string;
    dpRegistrationNumber?: string;
    bankName?: string;
    bankAddress?: string;
    isAssociatedWithSecuritiesMarket?: boolean;
    detailsOfPriorAssociation?: string;
    doesHoldPan?: boolean;
    panNumber?: string;
    isViolatedLaw?: boolean;
    createdAt?: Date;
    updatedAt?: Date;
}